backend/requirements.txt
flask
flask-cors
flask-jwt-extended
psycopg2-binary
python-dotenv
bcrypt